Reset to Cup
This is a complete reworking of the draw system where instead of resetting the terminal and rendering the full map for every movement, the cursor is moved to the appropriate map coordinates and the snake is drawn and erased where necessary. This saves a lot of time and processes, and prevents screen flicker.

Easter egg: press the 'end' key on the title screen.