Score Screen, New Game and Removed Timer
The clock has been replaced with a high score indicator that remains constant until a new game is started with a score higher than it.
Score screen has been added finally so the window doesn't just close when you get a game over.
Optimisations in tput cup printing, game creation, menus and in-terminal usage.
Use can now clear the high score from the menu. This is irreversible, but the file can be edited manually.